-machine.vim -ClearRegisters.vim +symlinks
Changed setup to be more streamlined, with no machine specific
configuration needed and everything automated. Defaulting folder to the
user profile/git/garbage, with the idea of it being easy to navigate to
my git folder where I put general repos. Symlinks for the vimrcs in
batch and shell script format (Untested).